Center the map initially
This commit eliminates the need to move the map to your location every time you open the page by assuming the first plane that's picked up is in the same general area as all the others which will be picked up, which is a reasonable assumption for most use cases. This might be preferable to continually setting the center based on the plane data because this method doesn't remove any freedom from the user to pan around the map manually.